improve AUTH handling of SMTP dissector.  Bug 8591 (https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8591)
From Uli Heilmeier

The current version of the SMTP dissector expects a 'AUTH LOGIN' mechanism without checking the mechanism.

When some other mechanism (like NTLM or PLAIN) is in use the decoding is wrong. Furthermore it is expected that the username is in a seperate packet. When the username is in the AUTH line the password is shown as smtp.auth.username and the username is not decoded.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@48910 f5534014-38df-0310-8fa8-9805f1628bb7