Let proto_tree_add_item() do the length checking for string lengths, so we thrown an exception if the lengths are too long.
For UCS-2/UTF-16 strings, do an extra length check to avoid overflows,
but throw ReportedBoundsError if the check fails, so we report them the
same way we report other too-large length errors.

Just use proto_tree_add_item() to put UCS-2/UTF-16 strings (assumed to
be UTF-16, not UCS-2; is that correct?) into the protocol tree.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@48913 f5534014-38df-0310-8fa8-9805f1628bb7