Follow-up to r49055 as noted on https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7359
The BER integer dissection routines take an hf_id, but that can be -1.
Only fetch the type (to check signedness) if hf_id >= 0, as otherwise this
causes a dissector bug. Default to signed if given no hf_id - I don't know
whether this should be unsigned or not, but the old behaviour was that
everything was signed so it's not a regression at least.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@49101 f5534014-38df-0310-8fa8-9805f1628bb7