As suggested by Evan:
Point users to https://code.google.com/p/pyreshark/ as an alternative to the
current Python support.

While we're in there, make it harder to enable Python: change it from
"--with-python" to "--with-broken-python" just to prevent people from enabling
unless they really mean it (are going to work on fixing it).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@49138 f5534014-38df-0310-8fa8-9805f1628bb7