Have register_param_stat() add menu items, rather than having the menu items hardcoded into ui/gtk/main_menubar.c
Get rid of more stuff implying that the code for adding menu items is
used only by Lua, as it's also used by register_param_stat().

Add the items to the menu list sorted by the menu item name shown to the
user.

Get rid of some const declarations that are lies.

Use the label given for a menu item as the label in the menu, rather
than parsing it from a token.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@49552 f5534014-38df-0310-8fa8-9805f1628bb7