(Finally!) check in part of Didier's patch to fix https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3290 (TRY_TO_FAKE_THIS_ITEM disables bounds errors):
Before calling TRY_TO_FAKE_THIS_ITEM() check if the length given (or, in
the case of FT_UINT_{STRING,BYTES}, the length we retrieve from the TVB)
exceeds what's left in the TVB.

Do this only for proto_tree_add_item() for now (it's the most commonly used
and thus the biggest trouble maker in this area).

Similar changes for other APIs will come later (if nothing blows up).  Despite
the fuzz failures this bug has caused I'm not sure about back-porting it...


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@49644 f5534014-38df-0310-8fa8-9805f1628bb7