Get out of global variable hell and other optimizations.
I'm stil not sure the copying of the packet_info structure is necessary (or what the consequences are), but at least it's not global anymore.

col_set_fence may be needed to prevent overwriting by PPP dissector,  Caching COL_INFO (or the other columns) into it's own string was way more work than necessary and could lead to no column info on a malformed packet.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@49689 f5534014-38df-0310-8fa8-9805f1628bb7