Add a new environment variable to cause abort()s if we add too many items to the tree (to separate this case from the generic DISSECTOR_BUG case).
Enable this environment variable when fuzz testing.

Enable the 3rd (without tree but with a read filter) check (added in r49643)
when testing capture files but not when fuzz testing--not sure if we want to
add even more to the fuzzbot's work load now (OTOH I've been running it for
a while and it hasn't buried me in bugs).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@49784 f5534014-38df-0310-8fa8-9805f1628bb7