Prevent attempts to allocate huge amounts of memory by imposing an artificial 16MB limit on blocks.
Do some sanity checks when reading options, to make sure we don't read
past the end of the block.

Make some variables unsigned so as not to get inappropriate
sign-extension (which, in practice, should never happen due to the 16MB
block size limit, although if the limit is raised above 2^31-1, the
limit won't protect you).

Fixes bug 8752.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@49833 f5534014-38df-0310-8fa8-9805f1628bb7