Fix for bug 8826.
The global variable 'address_item' was not always being set to NULL
before the dissection of a new frame.  Do more to set it in the general
case, and in the particular case of an MTSBindResult_PDU, which was
triggering an error in the supplied capture.

There may be other entry points where this still safely isn't being
set.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50083 f5534014-38df-0310-8fa8-9805f1628bb7