get_ax25_name() is *NOT* the right routine to use to translate a human-readable representing an AX.25 subaddress (e.g. "KA9Q-01") into the binary form of an AX.25 address, because what it does is translate the binary form of an AX.25 address to the human-readable form!
We currently have no routine that does the right thing and, even if we
did, given that some bits in the AX.25 subaddress format are used for
purposes other than representing the call sign and substation ID, so the
matching routines for AX.25 addresses need to ignore certain bits.

For now, we just remove the call to get_ax25_name() (which squelches the
pointer-signedness warning that made me look at this code, and find the
problem, in the first place) and replace it with a comment discussing
the problem and a failure.

The other pointer-signedness warning brought up a question of what to do
with G_REGEX_RAW in the g_regex_match_full() call; it didn't bring up an
immediately obvious *answer*, so we throw a cast at the warning and add
another comment.  (We fix up alignment while we're at it.)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50106 f5534014-38df-0310-8fa8-9805f1628bb7