Make TCP segment length a generated field instead of a hidden field.
We should deprecate the use of hidden fields, at least for fields that arei
useful in filters. To make it easier for users to discover and use the fields.

Change the highly useful field for TCP segment payload length from
being a hidden field to be a generated field instead.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50112 f5534014-38df-0310-8fa8-9805f1628bb7