From Dirk via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8850 :
The attached patch fixes the integer type of the WCCP identity mask value.
This is a bitmask which should be printed as hex, it doesn't make sense to
print it as an IPv4 address. See
http://tools.ietf.org/id/draft-wilson-wrec-wccp-v2-01.txt section 5.7.7 and
the attached capture file as an example.

The current draft http://tools.ietf.org/html/draft-mclaggan-wccp-v2rev1-00#section-6.15
doesn't mention "mask" in the names of the field any more, but the description
still describes them as mask values.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50211 f5534014-38df-0310-8fa8-9805f1628bb7