Make expert items filterable and a little cleanup.
Remove NULL tree check at start of dissection to ensure expert info is populated.

Overall I think the logic could be a little cleaner (length checking and its tie to expert info seems excessive), but I'm not familiar enough to be comfortable making the changes.

This dissector appears to be vulnerable to DOS attacks through its unsigned 16-bit length fields (just from static inspection).  Having the length be a signed 32-bit value even if the length in the protocol field is an unsigned 16-bit value could simplify some of the logic.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50310 f5534014-38df-0310-8fa8-9805f1628bb7