Get rid of an unused variable.
Don't nest g_strconcat() calls: g_strconcat(a, g_strconcat(b, c, NULL), NULL)
is equivalent to g_strconcat(a, b, c, NULL).  (And g_strconcat(b, c) is
incorrect - you need a NULL at the end of the list.)

Checking whether a pointer is "> 0" is useful only in platform-dependent
situations or if you're doing a really greasy hack such as stuffing a
flag into the uppermost bit of the pointer; the test should just check
whether the pointer is null or not.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50500 f5534014-38df-0310-8fa8-9805f1628bb7