Fix the problem reported in https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8908#c3 (the last remaining problem in that bug report):
When we expand a tree, check if any of its children need to be expanded too
(because the user left them expanded when s/he closed the parent).

In the process, rename expand_finfos() to check_expand_children() since the
function expands any children (subtrees) which should be expanded.  Remove
one of that function's parameters too: just reference the global preference
that controls whether scrolling should be done or not.

Replace some tabs with spaces (for consistency).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50516 f5534014-38df-0310-8fa8-9805f1628bb7