Use PIE_CFLAGS when building qt-shark, since otherwise I get errors like:
/usr/include/qt5/QtCore/qglobal.h:1079:4: error: #error "You must build your
code with position independent code if Qt was built with -reduce-relocations. "
"Compile your code with -fPIC or -fPIE."

There is almost certainly a better way to do this: I'm cheating here because I
don't know enough about autofoo, and it seems to work. I suspect we should be
checking PIE for the C++ compiler separately (since presumably they're not
guaranteed to be related like gcc/g++ are). I also suspect we should only be
building with it in the first place if Qt really needs it, though I have no idea
how to check that. The qt docs I've found suggest that qt5's cmake module has a
flag for it, but autofoo is never mentioned...



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50645 f5534014-38df-0310-8fa8-9805f1628bb7