From  Cal Turney: Unrecognized preferences and color filters created in proprietary or older versions are discarded when saved.
If the user attempts to save the preferences or colorfilters file, a popup is displayed that warns that unrecognized prefs or color filters have been detected and will be discarded if the save operation is allowed to proceed. In the case of Preferences, the popup message includes the version at which the file was last saved.  A "Continue without Saving" button is provided so that the user can save the profile under a different name.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8942

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50690 f5534014-38df-0310-8fa8-9805f1628bb7