Explain what the big loop in update_properties_all() is doing, and what's done after that.
If we want to set the all-interfaces capture filter string, just set it,
don't add anything to the drop-down list for it.

If, after we've succeeded starting a capture, all active interfaces have
the same capture filter, *do* add that filter to the all-interfaces
recent capture filters list.

Also, free g_strduped capture filter strings when we're done with them.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@50986 f5534014-38df-0310-8fa8-9805f1628bb7