From Mark Phillips via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9038 802.11 HT Extended Capabilities B10 decode incorrect
#Backport 1.10 and 1.8

Within the 802.11 HT Capabilities IE the "HT Extended Capabilities" B10's name is incorrect.

It is currently titled "High Throughput". This is incorrect/missleading. It should be something like "High Throughput Control (+HTC)". The field indicates if the "High Throughput Control (HTC)" field is supported - not if HT itself is supported.

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@51339 f5534014-38df-0310-8fa8-9805f1628bb7