Someone is bound to miss these expert info items, so add back expert info items for ack and urgent pointer when their respective flags are not set; however, don't indicate "Broken TCP" and only add them as Notes rather than as Warnings as was previously the case prior to r51356 since nowhere does it indicate these fields must be zero when their respective bits aren't set.
Speaking of r51356, a clarification to its commit message is in order: Initially my intention was to only add the urgent pointer field when the URG bit was set; however, I then noticed that the acknowledgment number field was always being added irrespective of the ACK bit.  Had I made the change as I originally intended, it would have introduced an inconsistency.  After some deliberation, I opted for consistency, but botched the commit message.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@51431 f5534014-38df-0310-8fa8-9805f1628bb7