Fully deprecate read filter (-R) without two-pass (-2). It does exactly the same thing as the display filter (-Y) in that case except with more confusing semantics.
This also lets us fix -c in the single-pass case to unconditionally count
packets. This isn't the old behaviour (which counted them only if they passed
the read filter) but is more consistent with two-pass mode where they are
counted even if they pass the display filter, since they are counted on the
first pass and the display filter is applied on the second pass.

Anyone who wants to use -c to limit packet count conditionally on them passing a
filter should use it in tandem with -2 and -R: the read filter is applied on the
first pass before the count.

Fixes https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9048


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@51556 f5534014-38df-0310-8fa8-9805f1628bb7