Respect the layout preference (just the horizontal/vertical splitting of the panes, not yet which data goes in which pane).
You currently have to restart for the change to take effect. I also had to move
the preferences-reading (and thus the epan_init) call to before the main window
instantiation so that the preference was actually available for use. This means
that the progress-bar doesn't show up for very long any more, since most of the
work is already done by the time it appears.

I'm sure there is a way to delay just the layout setup until later, but I kept
getting slot errors I couldn't figure out...


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@51647 f5534014-38df-0310-8fa8-9805f1628bb7