Have ChmodBPF kludge^Wcreate as many BPF devices as possible at startup.
The welcome screen in the Qt port runs "dumpcap -S" to draw sparklines.
On OS X this means that it holds open a BPF device for each interface.
Trying to capture using another instance of Wireshark (or tcpdump, or
tshark, or...) will trigger the creation of an additional BPF device but
we won't have permission to use it. Forcing device creation at startup
works around this.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@52227 f5534014-38df-0310-8fa8-9805f1628bb7